+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   **terra-form**: error messages have been added for the following form field validators:
     -   validator **required**,
+    -   validator **pattern**,
     -   validator **minLength**,
     -   validator **maxLength**,
     -   validator **email**,